Creates utilities and grid partials
To isolate the new styles from existing ones, I created a new partial for
variables, mixins, etc and called it utilities as well as a new partial for
the Zen grid system.  This allowed me to fix the header layout as it was not
properly making use of the grid.

mario h.
node/59666