Adjusts left-col top margin to align with other regions
Homepage spacing of left column does not match that of other regions
after a block no longer needed was removed.  Adjusted top margin of
left-region to perfectly align with other regions and keep a consistency
in spacing between the header area and the begining of left and right
regions.

mario h.
node/