Themes main navigation
Navigation background was achieved with an image in the past.  This was causing
issues when users zoomed in.  The image was not able to resize with the rest of
the site.  I refactored the navigation by changing the markup to be able to use
css for the nav background.

mario h.
node/60582