Fixes padding and dashed border on sidebar
Based on the type and number of blocks on the sidebar, the padding and
dashed border found in between each element needs to be managed to
ensure there is always padding in between elements and the dashed
border never appears after the last media block on the sidebar.

mario h.
node/60525