Fixes pager navigational arrows
Isse was reported that left navigation arrow was not displaying in Safari-based
devices.  This was probably caused by Transform CSS3 property used on left arrow
which used the right-pointing arrow and rotates it 180deg to make it point to
the left.
Created new image which combined both, right and left arrows and this corrected
the problem.

mario h.
node/62503