Themes footer so it always remains at bottom
On pages where content was too short to fill the entire page, the
footer was being pulled up leaving a large amount of empty space
below the footer.  Corrected by writing rules that always push the
footer to the bottom of the site regardless of amount of content
on page.

mario h.
node/63311