Themes Slac Science image and overlay
In addition to theming the slac science block's image and its overlay, I
styled other element's font size and alignments to complete mobile homepage.
Also removed JS which was originally used to handle resizing of images.  This
was causing a bug when browser was resized.  Image resizing will be handle with
CSS.

mario h.
node/63561