63790-Styles table captions
Styled table captions so they are centered-align on large screens.
On mobile and tablet the center-alignment does not work due to
responsive design styles applied to tables.  The only way to be
able to center align captions in mobile woudl be to set tables to
display as "table", this however could create issues on responsive
tables.

mario h