63799--Fixes empty header width on my flea market listing page
When content does not exist on my flea market listing page the empty
text shown in the header needed to inherit the same styles as the
header text when items do exist.  Refactored code and included the
empty-view class provided by views for when no items exist to allow
same styles on either situation.

mario h.