61377--time field format and grid adjustments
https://project.mediacurrent.com/slactoday/node/61377

The time field in the flea market table needed to not be inline with
the date field.  I applied display: block to the time field to allow it
move to its own line.
Also made minor grid adjustments to the flea market page to allow for
better table content display.

mario h.