63311--Keep footer at bottom of page
https://project.mediacurrent.com/slactoday/node/63311

Despite the fact I had tried to get this figured it out in previous
ticket, I ran into side effects that were causing other layout issues
with some pages.  I took another look at this issue and I believe
this is a better approach.
I modified page.tpl.php to move the footer container out of the l-page
container.  Doing this allowed me get the footer to remain at the bottom
of the page when there isn't enough content to cover the entire page while
moving the footer down if there is plenty of content on the page.

mario h.