64092--Set fixed height to slider on mobile
https://project.mediacurrent.com/slactoday/node/64092

Inconsistencies in the length of slider titles were causing issues
on mobile as the height of the title wrappers would change based on
the title length.  This was presenting UI issues.
Corrected by assigning a fixed height to the title wrapper to ensure
its background would remain consistent at the same height for all
slides.  Allowed for titles to wrap to up to 3 lines.  These changes
were reset on tablet and higher as on those devices we want the
title and teaser to expand freely.

mario h.