62788--Remove calendar module default styles
https://project.mediacurrent.com/slactoday/node/62788

Issues with cell height issues were being identified in the calendar's
day view.  These issues were random and hard to track.
Was able identify the source of the styles creating the height issues
and they happen to be styles that are part of the calendar module.

Because these are dandom issues it would not be practical to address them
on a individual basis.  We opted to completely remove all styles by
calendar module that affect the height of cells on day view.  Main reason
for doing this is that we are not using those styles by default.

mario h.