64331--Update gold color shade and font weight
https://project.mediacurrent.com/slactoday/node/64331

The current gold shade on block borders in homepage was showing
a little too dark and the same withe the font weight on the
upcoming events block in hoempage.
Client provided a lighter shade gold for the borders and requested
to turn font weight to normal for the date.
Made both of those changes.

mario h.