Adds flag link pane for anonymous users
* This pane is only displayed to anonymous
users and redirect them to the user login form
and sets the destination parameter in the url
so they are redirected back to the page where
they clicked the link.

Be aware that this will not resolve the issue with
webauth module but it does provide a link and force
authentication.

https://project.mediacurrent.com/slactoday/node/61485