64837--Fix text overlap in taxonimy listing pages
https://project.mediacurrent.com/slactoday/node/64837

Title of term listing pages was collapsing with the title of the
first node on the page.  This was due to css rules that were in
place that altered the page title position if a user is logged and
primary tabs are available.  However, the issue was not evident on
users who despite being logged in did not see the primary tabs due
to their role/access rights.

Corrected original rules by ensure the position of the page title
is not based on whether the user is logged in by instead whether
the primary tabs are present.  Fixing this issue actual broke the
position of the RSS feed icon.  Corrected this by applying the
same logic to its position.

mario h.