Exports Colorbox module settings fixing upload bug
The main issue being addressed in this commit is
the media file upload bug. The colorbox jQuery
library file was being loaded on the node form
even though the add/node/* was in the list of
pages to avoid. However the media module calls
ajax paths of media/ajax which once triggered
would load the colorbox library and prevent the
add another media button from being generated.

* Exports default colorbox settings
* Adds media/ajax and media/ajax/* to list of
paths to avoid.

https://project.mediacurrent.com/slactoday/node/64327