64965--Left align page header
https://project.mediacurrent.com/slactoday/node/64965

On tablet and beyond, the page header for taxonomy listing pages was
showing a left padding of 20px originated from the grid configuration
applied to the header. Corrected by removing the left padding allowing
the header to be left align with the rest of the content on the page.

mario h.