64911--Set background on expanded mobile nav
https://project.mediacurrent.com/slactoday/node/64911

When mobile nav is expanded the background is gray but it should
be white.  This issue was only visible on homepage where the page's
background is gray.  Corrected by changing the background to white
from transparent.  Change only affects homepage and only when
mobile nav is expanded/open.

mario h.