65150--Set new aspect ratio to featured news images
https://project.mediacurrent.com/slactoday/node/65150

Images for featured news on homepage slider and slac science were
showing in a 6:9 aspect ratio.  The request was for the images to
be displayed in a 3:2 aspect ratio.  Using an online aspect ratio
calculator I changed the images styles used by the featured news
images to use the new aspect ratio.
Flushed the image styles cache by running drush image-flush for the
respective image styles.

mario h.