65412--Fix videos captions
https://project.mediacurrent.com/slactoday/node/65412

Captions on videos was showing at 50% the width of the video.  This was
due to the fact that captions are shared between videos and images
and depending on where those media elements appear their widths vary.
Corrected by providing more context to captions.  For example, if the
caption was for a video the width would reflect the max width of the
video based on the image style used by the video.  Same goes for
images.

mario h.