65465--Fix IE11 navigation bug
https://project.mediacurrent.com/slactoday/node/65465

In IE10 and 11, when hovering over the main navigation or a link
in active state, a 1px bottom border would show which was originated
from the navigation background.  Fixed it by reducing the height of
the navigation by 1px therefore removing the border.
Also adjusted right margin on list items in navigation to remove
side border.  This is a bug that happens when displaying list items
inline.

mario h.