Adds image_info_cach and updates features
* This module generates the configured image
styles.
* Imageinfo Cache allows you to set certain
styles to be auto generated on entity save|update.
It also provides a drush command to do the same
in case you need to flush all of the image style
files. If you do flush all of the image styles
or at least the teaser style that is used in the
Mailchimp campaigns, it would break all of the
images in the email. But that can be fixed by
running the drush command that Imageinfo Cache
provides.

https://project.mediacurrent.com/slactoday/node/64785#comment-884985